endpoints for category, product and transaction DONE; to be tested

+++ src/main/java/tqs/marketplace/repositories/CategoryRepository.java
@@ -7,5 +7,5 @@ import tqs.marketplace.entities.Category;
 @Repository
 public interface CategoryRepository extends CrudRepository<Category, Long> {
     Category findByName(String name);
-    Category findById(int id);
+    Category findById(long id);
 }
